+++ example/lib/api/response_mappers.dart
@@ -1,6 +1,7 @@
-import 'package:example/api/models/errors/error_response_model.dart';
-import 'package:example/api/models/users_response_model.dart';
 import 'package:dio/dio.dart';
+import 'package:example/api/models/users_response_model.dart';
+
+import 'models/errors/error_response_model.dart';
 
 UsersResponseModel users(Response response) {
   if (response.statusCode == 200) {