+++ lib/src/exceptions/network_connection_exception.dart
@@ -3,7 +3,7 @@ import 'package:dio/dio.dart';
 class NetworkConnectionException implements Exception {
   NetworkConnectionException(this.error);
 
-  final DioError error;
+  final DioError? error;
 
   @override
   String toString() {