+++ test/api_client_test.dart
@@ -1,19 +1,20 @@
 import 'package:dash_kit_network/dash_kit_network.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
+import 'api_client_test.mocks.dart';
 import 'api_client_test_utils.dart';
-import 'mocks/mock_dio.dart';
-import 'mocks/mock_token_storage.dart';
 import 'test_components/test_api_client.dart';
 import 'test_components/test_refresh_tokens_delegate.dart';
 
+@GenerateMocks([TokenStorage, Dio])
 void main() {
-  Dio dio;
-  BaseOptions dioBaseOptions;
-  TestApiClient apiClient;
-  TestRefreshTokensDelegate delegate;
-  TokenStorage tokenStorage;
+  late Dio dio;
+  late BaseOptions dioBaseOptions;
+  late TestApiClient apiClient;
+  late TestRefreshTokensDelegate delegate;
+  late TokenStorage tokenStorage;
 
   setUp(() {
     dio = MockDio();
@@ -24,8 +25,8 @@ void main() {
   });
 
   test('No tokens exists', () async {
-    stubAccessToken(tokenStorage, null);
-    stubRefreshToken(tokenStorage, null);
+    stubAccessToken(tokenStorage, '');
+    stubRefreshToken(tokenStorage, '');
     stubDioOptions(dio, dioBaseOptions);
 
     apiClient = TestApiClient(dio, delegate);
@@ -34,22 +35,26 @@ void main() {
       return Future.error(DioError(
         response: Response(
           statusCode: 401,
-          request: RequestOptions(path: ''),
+          requestOptions: RequestOptions(path: ''),
         ),
+        requestOptions: RequestOptions(path: ''),
       ));
     });
 
     onRefreshRequestAnswer(dio, () {
       return Future.error(DioError(
-          response: Response(
-        statusCode: 400,
-        request: RequestOptions(path: ''),
-      )));
+        response: Response(
+          statusCode: 400,
+          requestOptions: RequestOptions(path: ''),
+        ),
+        requestOptions: RequestOptions(path: ''),
+      ));
     });
 
     final usersRequest = apiClient.get(
       path: 'users',
       isAuthorisedRequest: true,
+      responseMapper: (response) => response,
     );
 
     bool isRequestFailed = false;
@@ -63,7 +68,7 @@ void main() {
 
     verifyInOrder([
       dio.options,
-      userRequest(dio, accessToken: null),
+      userRequest(dio, accessToken: ''),
       refreshTokensRequest(dio),
     ]);
 
@@ -85,16 +90,16 @@ void main() {
       return Future.value(Response(
         statusCode: 200,
         data: ['John', 'Mary'],
-        request: RequestOptions(path: ''),
+        requestOptions: RequestOptions(path: ''),
       ));
     });
 
     final users = await apiClient.get(
-      path: 'users',
-      isAuthorisedRequest: true,
-    );
+        path: 'users',
+        isAuthorisedRequest: true,
+        responseMapper: (response) => response);
 
-    expect(users, ['John', 'Mary']);
+    expect(users.data, ['John', 'Mary']);
 
     verifyInOrder([
       dio.options,
@@ -117,15 +122,18 @@ void main() {
 
     onUserRequestAnswer(dio, () {
       return Future.error(DioError(
-          response: Response(
-        statusCode: 403,
-        request: RequestOptions(path: ''),
-      )));
+        response: Response(
+          statusCode: 403,
+          requestOptions: RequestOptions(path: ''),
+        ),
+        requestOptions: RequestOptions(path: ''),
+      ));
     });
 
     final usersRequest = apiClient.get(
       path: 'users',
       isAuthorisedRequest: true,
+      responseMapper: (response) => response,
     );
 
     bool isRequestFailed = false;
@@ -161,16 +169,18 @@ void main() {
       if (counter < 1) {
         counter++;
         return Future.error(DioError(
-            response: Response(
-          statusCode: 401,
-          request: RequestOptions(path: ''),
-        )));
+          response: Response(
+            statusCode: 401,
+            requestOptions: RequestOptions(path: ''),
+          ),
+          requestOptions: RequestOptions(path: ''),
+        ));
       }
 
       return Future.value(Response(
         statusCode: 200,
         data: ['John', 'Mary'],
-        request: RequestOptions(path: ''),
+        requestOptions: RequestOptions(path: ''),
       ));
     });
 
@@ -181,18 +191,19 @@ void main() {
           'access_token': '<refreshed_access_token>',
           'refresh_token': '<refreshed_refresh_token>',
         },
-        request: RequestOptions(path: ''),
+        requestOptions: RequestOptions(path: ''),
       ));
     });
 
     final usersRequest = apiClient.get(
       path: 'users',
       isAuthorisedRequest: true,
+      responseMapper: (response) => response,
     );
 
     final users = await usersRequest;
 
-    expect(users, ['John', 'Mary']);
+    expect(users.data, ['John', 'Mary']);
 
     verifyInOrder([
       dio.options,
@@ -228,23 +239,28 @@ void main() {
 
     onUserRequestAnswer(dio, () {
       return Future.error(DioError(
-          response: Response(
-        statusCode: 401,
-        request: RequestOptions(path: ''),
-      )));
+        response: Response(
+          statusCode: 401,
+          requestOptions: RequestOptions(path: ''),
+        ),
+        requestOptions: RequestOptions(path: ''),
+      ));
     });
 
     when(dio.post('refresh_tokens')).thenAnswer(
       (_) => Future.error(DioError(
-          response: Response(
-        statusCode: 401,
-        request: RequestOptions(path: ''),
-      ))),
+        response: Response(
+          statusCode: 401,
+          requestOptions: RequestOptions(path: ''),
+        ),
+        requestOptions: RequestOptions(path: ''),
+      )),
     );
 
     final usersRequest = apiClient.get(
       path: 'users',
       isAuthorisedRequest: true,
+      responseMapper: (response) => response,
     );
 
     try {