+++ test/mocks/mock_dio.dart
@@ -1,4 +1,0 @@
-import 'package:dash_kit_network/dash_kit_network.dart';
-import 'package:mockito/mockito.dart';
-
-class MockDio extends Mock implements Dio {}